rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Local RAG system for Claude Code conversation and codebase search"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0,<1.0",
    "mcp[cli]>=1.0.0,<2.0",
    "pyyaml>=6.0",
]

[project.scripts]
rag-server = "src.server:main"
rag-indexer = "src.indexer:main"
rag-watcher = "src.watcher:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
